feat: init i18n in site
Initialized localization in sugaming-site.
