+++ apps/sugaming-site/app/[locale]/page.module.scss
