+++ apps/sugaming-site/app/[locale]/page.tsx
@@ -3,9 +3,12 @@
 import { useState } from 'react';
 import { ApiClient } from '@sugaming/sugaming-api-client/client';
 import { getAuth, login } from '@sugaming/sugaming-api-client/next';
+import { useTranslations } from 'next-intl';
 
 export default function Index() {
+  const t = useTranslations('Index');
   const [userInfo, setUserInfo] = useState<ApiClient.UserDto>();
+
   const handleClick = async () => {
     await login('gosho@losho.com', 'GoshoLoshoTestPassword');
     const res = await ApiClient.UsersApiService.usersControllerGetProfile({
@@ -16,7 +19,7 @@ export default function Index() {
 
   return (
     <>
-      <h1>Page</h1>
+      <h1>{t('title')}</h1>
 
       <button type="button" onClick={handleClick}>
         execute server action
