--- apps/sugaming-site/app/layout.tsx
@@ -1,18 +1,0 @@
-import './global.css';
-
-export const metadata = {
-  title: 'Welcome to sugaming-site',
-  description: 'Generated by create-nx-workspace',
-};
-
-export default function RootLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return (
-    <html lang="en">
-      <body>{children}</body>
-    </html>
-  );
-}
