+++ apps/sugaming-site/i18n/bg.json
@@ -1,0 +1,5 @@
+{
+  "Index": {
+    "title": "Тест заглавие"
+  }
+}
