+++ apps/sugaming-site/i18n/en.json
@@ -1,0 +1,5 @@
+{
+  "Index": {
+    "title": "Test title"
+  }
+}
