+++ package.json
@@ -21,6 +21,7 @@
     "ms": "^2.1.3",
     "nestjs-i18n": "^10.3.7",
     "next": "^14.0.2",
+    "next-intl": "^3.1.2",
     "passport": "^0.6.0",
     "passport-jwt": "^4.0.1",
     "passport-local": "^1.0.0",
